+++ opentelemetry-sdk/src/attributes/mod.rs
@@ -1,3 +1,4 @@
 mod set;
 
+#[cfg(feature = "metrics")]
 pub use set::AttributeSet;
